+++ common/src/main/java/org/astraea/common/balancer/algorithms/AlgorithmConfig.java
@@ -18,10 +18,16 @@ package org.astraea.common.balancer.algorithms;
 
 import java.time.Duration;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.function.BiPredicate;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterBean;
+import org.astraea.common.balancer.Balancer;
 import org.astraea.common.cost.ClusterCost;
 import org.astraea.common.cost.Configuration;
 import org.astraea.common.cost.HasClusterCost;
@@ -30,6 +36,10 @@ import org.astraea.common.cost.MoveCost;
 
 /** The generic algorithm parameter for resolving the Kafka rebalance problem. */
 public interface AlgorithmConfig {
+
+  static Builder builder() {
+    return new Builder();
+  }
 
   /** @return the cluster cost function for this problem. */
   HasClusterCost clusterCostFunction();
@@ -51,4 +61,163 @@ public interface AlgorithmConfig {
 
   /** @return the algorithm implementation specific parameters */
   Configuration algorithmConfig();
+
+  class Builder {
+
+    private HasClusterCost clusterCostFunction;
+    private List<HasMoveCost> moveCostFunction = List.of(HasMoveCost.EMPTY);
+    private BiPredicate<ClusterCost, ClusterCost> clusterConstraint =
+        (before, after) -> after.value() < before.value();
+    private Predicate<List<MoveCost>> movementConstraint = ignore -> true;
+    private int searchLimit = Integer.MAX_VALUE;
+    private Duration executionTime = Duration.ofSeconds(3);
+    private Supplier<ClusterBean> metricSource = () -> ClusterBean.EMPTY;
+    private Configuration algorithmConfig = Configuration.of(Map.of());
+
+    /**
+     * Specify the cluster cost function to use. It implemented specific logic to evaluate if a
+     * rebalance plan is worth using at certain performance/resource usage aspect
+     *
+     * @param costFunction the cost function for evaluating potential rebalance plan.
+     * @return this
+     */
+    public Builder clusterCost(HasClusterCost costFunction) {
+      this.clusterCostFunction = Objects.requireNonNull(costFunction);
+      return this;
+    }
+
+    /**
+     * Specify the movement cost function to use. It implemented specific logic to evaluate the
+     * performance/resource impact against the cluster if we adopt a given rebalance plan.
+     *
+     * @param costFunction the cost function for evaluating the impact of a rebalance plan to a
+     *     cluster.
+     * @return this
+     */
+    public Builder moveCost(List<HasMoveCost> costFunction) {
+      this.moveCostFunction = Objects.requireNonNull(costFunction);
+      return this;
+    }
+
+    /**
+     * Specify the cluster cost constraint for any rebalance plan.
+     *
+     * @param clusterConstraint a {@link BiPredicate} to determine if the rebalance result is
+     *     acceptable(in terms of performance/resource consideration). The first argument is the
+     *     {@link ClusterCost} of current cluster, and the second argument is the {@link
+     *     ClusterCost} of the proposed new cluster.
+     * @return this
+     */
+    public Builder clusterConstraint(BiPredicate<ClusterCost, ClusterCost> clusterConstraint) {
+      this.clusterConstraint = clusterConstraint;
+      return this;
+    }
+
+    /**
+     * Specify the movement cost constraint for any rebalance plan.
+     *
+     * @param moveConstraint a {@link Predicate} to determine if the rebalance result is
+     *     acceptable(in terms of the ongoing cost caused by execute this rebalance plan).
+     * @return this
+     */
+    public Builder movementConstraint(Predicate<List<MoveCost>> moveConstraint) {
+      this.movementConstraint = Objects.requireNonNull(moveConstraint);
+      return this;
+    }
+
+    /**
+     * Specify the maximum number of rebalance plans for evaluation. A higher number means searching
+     * & evaluating more potential rebalance plans, which might lead to longer execution time.
+     *
+     * @deprecated The meaning of this parameter might change from algorithm to algorithm. Should
+     *     use {@link Builder#config} instead.
+     * @param limit the maximum number of rebalance plan for evaluation.
+     * @return this
+     */
+    @Deprecated
+    public Builder limit(int limit) {
+      // TODO: get rid of this method. It proposes some kind of algorithm implementation details.
+      this.searchLimit = Utils.requirePositive(limit);
+      return this;
+    }
+
+    /**
+     * @param limit the execution time of searching best plan.
+     * @return this
+     */
+    public Builder limit(Duration limit) {
+      this.executionTime = limit;
+      return this;
+    }
+
+    /**
+     * Specify the source of bean metrics. The default supplier return {@link ClusterBean#EMPTY}
+     * only, which means any cost function that interacts with metrics won't work. To use a cost
+     * function with metrics requirement, one must specify the concrete bean metric source by
+     * invoking this method.
+     *
+     * @param metricSource a {@link Supplier} offers the newest {@link ClusterBean} of target
+     *     cluster
+     * @return this
+     */
+    public Builder metricSource(Supplier<ClusterBean> metricSource) {
+      this.metricSource = metricSource;
+      return this;
+    }
+
+    /**
+     * @param configuration for {@link Balancer}
+     * @return this
+     */
+    public Builder config(Configuration configuration) {
+      this.algorithmConfig = configuration;
+      return this;
+    }
+
+    public AlgorithmConfig build() {
+      return new AlgorithmConfig() {
+        @Override
+        public HasClusterCost clusterCostFunction() {
+          return clusterCostFunction;
+        }
+
+        @Override
+        public List<HasMoveCost> moveCostFunctions() {
+          return moveCostFunction;
+        }
+
+        @Override
+        public BiPredicate<ClusterCost, ClusterCost> clusterConstraint() {
+          return clusterConstraint;
+        }
+
+        @Override
+        public Predicate<List<MoveCost>> movementConstraint() {
+          return movementConstraint;
+        }
+
+        @Override
+        public Duration executionTime() {
+          return executionTime;
+        }
+
+        @Override
+        public Supplier<ClusterBean> metricSource() {
+          return metricSource;
+        }
+
+        private final Configuration prepared =
+            Configuration.of(
+                Stream.concat(
+                        algorithmConfig.entrySet().stream(),
+                        Stream.of(Map.entry("iteration", String.valueOf(searchLimit))))
+                    .collect(Collectors.toUnmodifiableMap(Map.Entry::getKey, Map.Entry::getValue)));
+
+        @Override
+        public Configuration algorithmConfig() {
+          return prepared;
+        }
+      };
+    }
+  }
 }
